requirements.txt
python-dotenv==0.19.2
jinja2==3.0.3
django-jinja==2.10.0

pytest==6.2.5
pytest-cov==3.0.0
pytest-dotenv==0.5.2
pytest-django==4.4.0
pytest-pythonpath==0.7.3
pytest-tldr==0.2.4
pytest-xdist==2.5.0
